fixes and cleanups.
We now build a bootable image. extlinux is the bootloader for qemu x86 since it
doesn't seem possible to install grub-pc unattended and get it to work.
We also use raw files instead of qcow2 because there seems to be a race condition
that causes mkfs.ext4 on a qemu-nbd mounted qcow2 file to get stuck in D+ state.